Split up exporter tests.
